+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## Edge
+## 10.0.0-beta7
 
-- Throw an error before attempting to tag `null` or `undefined`.
+- Throw an error before attempting to tag `null` or `undefined`
 - Warn if `Presenter::viewModel` returns `repo.state` directly, allow
   `viewModel` to return a function similar to an individual key/value
   binding.